Migrate src/shobjidl to 0.3
