+++ Cargo.toml
@@ -164,6 +164,8 @@ servprov = []
 setupapi = []
 shellapi = []
 shlobj = []
+shlobjidl = []
+shlobjidl_core = []
 shtypes = []
 spapidef = []
 sql = []
