+++ src/um/mod.rs
@@ -98,6 +98,8 @@ pub mod gl;
 #[cfg(feature = "setupapi")] pub mod setupapi;
 #[cfg(feature = "shellapi")] pub mod shellapi;
 #[cfg(feature = "shlobj")] pub mod shlobj;
+#[cfg(feature = "shobjidl")] pub mod shobjidl;
+#[cfg(feature = "shobjidl_core")] pub mod shobjidl_core;
 #[cfg(feature = "shtypes")] pub mod shtypes;
 #[cfg(feature = "spapidef")] pub mod spapidef;
 #[cfg(feature = "sql")] pub mod sql;
